refactor: split event_log_storage tests
